Move link reading to separate module
